+++ services/role.services.js
@@ -1,23 +1,25 @@
 import Role from "../models/role.model.js";
 
 
-export const createRole = async(data) =>{
-    return await Role.create(data);
+export const createRole = async(
+    {data, 
+    tenantId}) =>{
+    return await Role.create({...data, tenantId});
 }
 
-export const getRole = async() => {
-    return await Role.find().populate('permission')
+export const getRole = async(tenantId) => {
+    return await Role.find({tenantId}).populate('permission')
     ;
 }
 
-export const getRoleById = async(id) => {
-    return await Role.findById(id);
+export const getRoleById = async(id,tenantId) => {
+    return await Role.findById({_id:id,tenantId});
 };
 
-export const updateRole = async(id , data) => {
-    return await Role.findByIdAndUpdate(id, data)
+export const updateRole = async(id ,tenantId, data) => {
+    return await Role.findByIdAndUpdate({_id:id,tenantId},data, {new :true});
 };
 
-export const deleteRole = async(id) => {
-    return await Role.findByIdAndDelete(id)
+export const deleteRole = async(id,tenantId) => {
+    return await Role.findByIdAndDelete({_id:id, tenantId});
 }
